use common component for Catalog and ShareView for editing the request
